Minor fix in session handling

+++ UserException.php
@@ -279,18 +279,14 @@ class UserException extends \yii\base\UserException {
 
 				// Delete expired errors once per session
 				$session = Yii::$app->session;
-				if ($session && !Yii::$app->request->isConsoleRequest) {  //will get message if run in CLI: "session_set_cookie_params(): Cannot change session cookie parameters when headers already sent"
+				if ($session && !Yii::$app->request->isConsoleRequest && !Yii::$app->response->isSent) {  //will get message if run in CLI: "session_set_cookie_params(): Cannot change session cookie parameters when headers already sent"
 					try {
 						if (!$session->get('wsErrorsCleared')) {
 							$dbConn->createCommand("DELETE FROM `". $databaseTable ."` WHERE err_expire_days IS NOT NULL AND TO_DAYS(err_timestamp) + err_expire_days < TO_DAYS(NOW())")->execute();
 							$session->set('wsErrorsCleared', 1);
 						}
 					} catch (\Exception $e) {
-						if (preg_match("/session_start.*headers already sent/", $e->getMessage())) {  //in case of CLI programs we might end up with this error because output was already echoed way before we got to here where Yii2 tries to open the session, so ignore that error
-							// ignore exception
-						} else {
-							throw $e;
-						}
+						throw $e;
 					}
 				}
 			} else {
